+++ src/index.ts
@@ -7,4 +7,5 @@
 export { encode, ZonEncoder } from './encoder';
 export { decode, ZonDecoder, DecodeOptions } from './decoder';
 export { ZonDecodeError, ZonDecodeErrorDetails } from './exceptions';
+export * from './schema';
 export * as constants from './constants';